"ads2_m6_a1_1: 3/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS-2 assignments/m6/ADS-2Week-1/ADS-2Week-1/Solution.java
@@ -51,7 +51,7 @@ class PageRank {
 				for (int k : digraph.reverse().adj(i)) {
 					//if (k == i) {
 					//if (digraph.outdegree(j) > 0) {
-					rank += pageR[k] / (double)digraph.outdegree(k);
+					rank += pageR[k] / digraph.outdegree(k);
 					//	}
 					//}
 					//}
